% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank_to_bank.R
\name{bank_to_bank}
\alias{bank_to_bank}
\title{Generate a line segment from bank to bank}
\usage{
bank_to_bank(pt, banks, angle)
}
\arguments{
\item{pt}{A point within the channel.}

\item{banks}{The bankline of the channel.}

\item{angle}{The angle of the line segment, in radians.}
}
\value{
A line segment spanning from bank to bank.
}
\description{
Given a point within a channel, generate a line segment that goes from
bank to bank, for a specified angle.
}
