% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banks_to_centerline.R
\name{banks_to_centerline}
\alias{banks_to_centerline}
\title{Function to calculate a centerline from bankline polygon}
\usage{
banks_to_centerline(sf_banks)
}
\arguments{
\item{sf_banks}{\code{sf}, \code{sfc}, or \code{SpatVector} polygons object.}
}
\value{
A
}
\description{
Function to calculate a centerline from bankline polygon
}
