% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cross_sections.R
\name{calculate_cross_sections}
\alias{calculate_cross_sections}
\title{Calculate Cross Sections from Banklines}
\usage{
calculate_cross_sections(n, banks)
}
\arguments{
\item{n}{The number of cross sections to calculate.}

\item{banks}{An sf object with banklines.}
}
\value{
An sf object with cross sections.
}
\description{
Calculate cross sections from banklines that are roughly equally spaced
apart.
}
\details{
This function takes the definition of "cross section" relative
to a point in the channel to be the line segment intersecting the point
whose bank-to-bank segment width is the smallest. Note that this does not
imply that the cross section is unique, and in this case the cross section
is arbitrarily taken to be the one closest to a 0-degree angle --
although in almost all cases this should not be an issue.

To define the spacing of the cross sections, a centerline is
first calculated, and equally spaced points are sampled along the
centerline. Cross sections are calculated at these points. Note that
cross sections are not a necessary part of choosing cross section spacing,
but it is useful.
}
\note{
The function of bank-to-bank width for a given angle of the line
crossing through a specified point is riddled with local minima. This
function takes a remedial approach at mitigating this:
it first finds the minimum width on a grid of 10 angles between 0 and pi
(inclusive) (or 100 if the minimum is not unique at first), and then
optimizes the width function around the minimum found on the grid.
An improved algorithm would be appropriate in the rare case the region
containing the global minimum is missed.
}
