% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_xs.R
\name{calculate_xs}
\alias{calculate_xs}
\title{Generate cross sections from a channel}
\usage{
calculate_xs(banks, centerline, num)
}
\arguments{
\item{num}{Number of cross sections to take from the channel.}

\item{bankline, centerline}{Bankline and centerline}
}
\value{
A list of cross sections.
}
\description{
Generate cross sections from a channel
}
\examples{
calculate_xs(my_channel, num = 100)
}
