% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_width.R
\name{ch_width}
\alias{ch_width}
\alias{ch_width<-}
\title{Get and Set Channel Width}
\usage{
ch_width(cross_section)

ch_width(cross_section) <- value
}
\arguments{
\item{cross_section}{Cross section.}

\item{value}{Channel width in meters; single positive numeric.}
}
\value{
Cross section width; single numeric.
}
\description{
Get and Set Channel Width
}
\examples{
cs <- cross_section(3, grad = 0.01, d50 = 45, d84 = 90, roughness = 0.01)
ch_width(cs)
ch_width(cs) <- 10
cs
}
