% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{channel}
\alias{channel}
\title{Simple feature channel object containing a centerline, a cross section,
a gradient and banks}
\usage{
channel(channel)
}
\arguments{
\item{channel}{Shapefile file containing channel geometries}
}
\description{
Simple feature channel object containing a centerline, a cross section,
a gradient and banks
}
