% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{cross_section}
\alias{cross_section}
\title{Create a new channel cross section}
\usage{
cross_section(width, grad, d50, d84, roughness, rootdepth = 0)
}
\arguments{
\item{width}{Channel width in meters; single positive numeric.}

\item{grad}{Energy gradient of the stream channel; single positive numeric.}

\item{d50, d84}{Grain size distribution's 50th and 84th quantiles in
millimeters; single positive numerics.}

\item{roughness}{Manning's roughness; positive single numeric.}

\item{rootdepth}{Effective rooting depth for vegetation; single non-negative
numeric.}
}
\value{
A \code{"cross_section"} object.
}
\description{
Create a channel cross section by specifying its properties.
}
\details{
Examples of effective rooting depth for vegetation, \code{rootdepth}, are:
\itemize{
\item grassy banks, no trees / shrubs: \code{rootdepth = 0.35}.
\item 1 to 5\% tree / shrub cover: \code{rootdepth = 0.50}.
\item 5 to 50\% tree / shrub cover: \code{rootdepth = 0.90}.
\item more than 50\% tree / shrub cover: \code{rootdepth = 1.10}.
}
}
\note{
Width is assumed constant as a function of depth.
}
\examples{
cross_section(3, grad = 0.01, d50 = 45, d84 = 80, roughness = 0.01)
}
