% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_bankline.R
\docType{data}
\name{demo_bankline}
\alias{demo_bankline}
\title{Example bankline polygon}
\format{
An object of class \code{sfc_POLYGON} (inherits from \code{sfc}) of length 1.
}
\usage{
demo_bankline
}
\description{
Example bankline polygon
}
\keyword{datasets}
