% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sxc.R
\name{is_sxc}
\alias{is_sxc}
\title{Test if an object is a cross section set.}
\usage{
is_sxc(x)
}
\arguments{
\item{x}{An object.}
}
\value{
Logical; TRUE if the test passes.
}
\description{
Checks whether an object inherits the "sxc" class.
}
