% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_banks.R
\docType{data}
\name{my_banks}
\alias{my_banks}
\title{Example bankline}
\format{
An object of class \code{sfc_POLYGON} (inherits from \code{sfc}) of length 1.
}
\usage{
my_banks
}
\description{
Example bankline
}
\keyword{datasets}
