% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{new_channel}
\alias{new_channel}
\title{Constructor function for channel objects}
\usage{
new_channel(l, ..., class = character())
}
\arguments{
\item{l}{Object to stamp with the "channel" class.}

\item{...}{Attributes to add to the object.}

\item{class}{Subclass; optional.}
}
\value{
The object \code{l} with the attributes in \code{...} and the class \code{"channel"}
with subclass \code{class} (if provided).
}
\description{
Not intended for front end use; to make a channel object, use \code{channel()}.
}
