% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{new_cross_section}
\alias{new_cross_section}
\title{Constructor function for cross_section objects}
\usage{
new_cross_section(l, ..., class = character())
}
\arguments{
\item{l}{List containing the components of a cross_section object.}

\item{...}{Attributes to add to the object.}

\item{class}{If making a subclass, specify its name here.}
}
\value{
An object of class \code{"cross_section"}, although not necessarily with
valid properties.
}
\description{
Constructor function for cross_section objects
}
