% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sx_centerline.R
\name{sx_centerline}
\alias{sx_centerline}
\title{Creates or obtains the centerline of the channel object}
\usage{
sx_centerline(channel, centerline = NULL)
}
\arguments{
\item{channel}{channel object}

\item{centerline}{shapefile containing the centerline linestring geometries
If not supplied, the function will delineate a centerline based on the channel
object}
}
\description{
Creates or obtains the centerline of the channel object
}
