% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sx_centerline.R, R/sx_xs.R
\name{sx_centerline}
\alias{sx_centerline}
\alias{sx_centerline<-}
\alias{sx_centreline}
\alias{sx_centreline<-}
\alias{sx_xs}
\alias{sx_xs<-}
\title{Creates or obtains the centerline of the channel object}
\usage{
sx_centerline(channel)

sx_centerline(channel) <- value

sx_centreline(channel)

sx_centreline(channel) <- value

sx_xs(channel)

sx_xs(channel) <- value
}
\arguments{
\item{channel}{channel object}

\item{value}{Centerline to set in channel.}
}
\description{
Creates or obtains the centerline of the channel object. \code{sx_centreline()}
is an alias for \code{sx_centerline()}.

Creates or obtains the centerline of the channel object. \code{sx_centreline()}
is an alias for \code{sx_centerline()}.
}
