% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sx_channel.R
\name{sx_channel}
\alias{sx_channel}
\title{Simple feature channel object containing a centerline, a cross section,
a gradient and banks}
\usage{
sx_channel(channel, class = character())
}
\arguments{
\item{channel}{Shapefile file containing channel geometries}
}
\description{
Simple feature channel object containing a centerline, a cross section,
a gradient and banks
}
