% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{validate_cross_section}
\alias{validate_cross_section}
\title{Validator function for cross_section objects}
\usage{
validate_cross_section(cross_section)
}
\arguments{
\item{cross_section}{Object of class \code{"cross_section"}.}
}
\value{
Returns the original \code{cross_section} object. Note that this function is
intended to be run for its side effects: namely, throwing an error if the
cross_section is invalid.
}
\description{
Checks that an object of class \code{"cross_section"} has a valid
structure and valid cross section properties.
}
