% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_cross_section_functions.R
\name{xt_cross_section}
\alias{xt_cross_section}
\title{Make cross section}
\usage{
xt_cross_section(mat, x_lb, x_rb)
}
\description{
Make cross section
}
\examples{
my_xs <- xt_cross_section(coords, 1, 12)
plot_(my_xs)

xs_width(my_xs)
rb_height(my_xs)

bigger_xs <- widen_right(my_xs, 2)
plot_(bigger_xs, add = TRUE)

xs_width(bigger_xs)
rb_height(bigger_xs)

plot_(widen_right(bigger_xs, 3), add = TRUE)
plot_(widen_left(bigger_xs, 3), add = TRUE)
}
