% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt_sort.R
\name{xt_distance}
\alias{xt_distance}
\title{Calculate the distance between two individual cross sections}
\usage{
xt_distance(xs1, xs2, centerline)
}
\arguments{
\item{xs1, xs2}{Individual cross sections (line segments)}

\item{centerline}{A line segment representing the centerline of the channel.}
}
\value{
The distance between the two cross sections as traversed along
the channel centerline.
}
\description{
Calculate the distance between two individual cross sections
}
