% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt_features.R, R/xt_sxc.R
\name{xt_features}
\alias{xt_features}
\title{Cross Section Features}
\usage{
xt_features(
  ...,
  grad = NULL,
  d50 = NULL,
  d84 = NULL,
  roughness = NULL,
  rootdepth = NULL
)

xt_features(
  ...,
  grad = NULL,
  d50 = NULL,
  d84 = NULL,
  roughness = NULL,
  rootdepth = NULL
)
}
\description{
Makes a list of cross section features. Useful because it
indicates features that are useful for algorithms like gbem;
otherwise, there is no difference from a list.

Makes a list of cross section features. Useful because it
indicates features that are useful for algorithms like gbem;
otherwise, there is no difference from a list.
}
