% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt_generate_centerline.R
\name{xt_generate_centerline}
\alias{xt_generate_centerline}
\title{Function to calculate a centerline from bankline polygon}
\usage{
xt_generate_centerline(sf_banks)
}
\arguments{
\item{sf_banks}{\code{sf}, \code{sfc}, or \code{SpatVector} polygons object.}
}
\value{
A centerline.
}
\description{
Function to calculate a centerline from bankline polygon
}
\details{
Currently just calls \code{centerline::cnt_path_guess()}.
}
