% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt_sort.R
\name{xt_sort}
\alias{xt_sort}
\title{Sort cross sections from upstream to downstream.}
\usage{
xt_sort(sxc)
}
\arguments{
\item{sxc}{sxc object, not sx}
}
\value{
The same cross section object, but sorted from
upstream to downstream.
}
\description{
Sort cross sections from upstream to downstream.
}
