% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt_widen_by.R, R/xt_widen_times.R
\name{xt_widen_by}
\alias{xt_widen_by}
\alias{xt_widen_times}
\title{Widen cross sections}
\usage{
xt_widen_by(object, by)

xt_widen_times(object, times)
}
\arguments{
\item{object}{A cross section object.}

\item{by}{Amount to widen the channel by, using units in common with the
cross sectional units. Negative values will narrow cross sections. Either
a vector of length equal to the number of cross sections, or length 1.}

\item{times}{Multiplicative factor to widen the channel by. Values less than
1 will narrow cross sections. Either
a vector of length equal to the number of cross sections, or length 1.}
}
\description{
Widen or narrow a collection of cross sections. \code{xt_widen_by()} widens
(narrows) by unit length; \code{xt_widen_times()} widens (narrows) by
a multiplicative factor.
}
