% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt_width.R
\name{xt_width}
\alias{xt_width}
\title{Get Channel Width}
\usage{
xt_width(object)
}
\arguments{
\item{object}{Cross section.}
}
\value{
Cross section width; single numeric.
}
\description{
Get Channel Width
}
\examples{
xs <- xt_sxc(1:3)
xt_width(xs)
}
