% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt_width.R
\name{xt_width}
\alias{xt_width}
\title{Get and Set Channel Width}
\usage{
xt_width(cross_section)
}
\arguments{
\item{cross_section}{Cross section.}
}
\value{
Cross section width; single numeric.
}
\description{
Get and Set Channel Width
}
\examples{
cs <- cross_section(3, grad = 0.01, d50 = 45, d84 = 90, roughness = 0.01)
ch_width(cs)
ch_width(cs) <- 10
cs
}
