% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{channel}
\alias{channel}
\title{Make a new channel}
\usage{
channel(bankline, centerline = NULL)
}
\arguments{
\item{bankline}{Bankline \code{sf} polygon.}

\item{centerline}{An optional \code{sf} centerline object. \code{NULL} if not provided.}
}
\value{
A new channel object.
}
\description{
Make a new channel
}
\examples{
channel(my_banks)
}
