% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_channel.R
\docType{data}
\name{my_banks}
\alias{my_banks}
\alias{my_channel}
\title{Example bankline and corresponding channel object}
\format{
An object of class \code{sfc_POLYGON} (inherits from \code{sfc}) of length 1.

An object of class \code{channel} of length 2.
}
\usage{
my_banks

my_channel
}
\description{
Example bankline and corresponding channel object
}
\keyword{datasets}
