% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sx_centreline.R
\name{sx_centreline}
\alias{sx_centreline}
\title{Creates or obtains the centreline of the channel object}
\usage{
sx_centreline(channel, centerline = NULL)
}
\arguments{
\item{channel}{channel object}

\item{centerline}{shapefile containing the centerline linestring geometries
If not supplied, the function will delineate a centerline based on the channel
object}
}
\description{
Creates or obtains the centreline of the channel object
}
