% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt_sx.R, R/xt_sxc.R
\name{xt_sx}
\alias{xt_sx}
\alias{xt_sxc}
\title{Create channel cross sections}
\usage{
xt_sx(x, ...)

xt_sxc(x, ...)
}
\arguments{
\item{x}{Object to create cross sections out of. Can be a vector of
(positive) widths if the spatial orientation is not important;
can be line segments created using the sf package.}

\item{...}{For \code{xt_sx()}, optional named entries of cross section
characteristics other than the ones listed as arguments;
lengths are recycled to match the number of cross sections.
For \code{xt_sxc()}, optional arguments passed to \code{sf::st_sfc()} relating to the
geometrical properties of the cross sections, such as \code{crs}.}

\item{grad}{Energy gradient of the stream channel; single positive numeric.}

\item{d50, d84}{Grain size distribution's 50th and 84th quantiles in
millimeters; single positive numerics.}

\item{roughness}{Manning's roughness; positive single numeric.}

\item{rootdepth}{Effective rooting depth for vegetation; single non-negative
numeric.}
}
\value{
For \code{xt_sx()}, cross section objects with class "sx",
which is a special type of "sf" object. For \code{st_sxc()}, cross section
geometry objects with class "sxc", which is a special type of "sfc"
object.
}
\description{
Create channel cross sections.
Cross section geometries (line segments) are created with \code{xt_sxc()}.
Cross section geometries can be paired with cross section characteristics
like roughness and gradient with \code{xt_sx()}.
}
\details{
Examples of effective rooting depth for vegetation, \code{rootdepth}, are:
\itemize{
\item grassy banks, no trees / shrubs: \code{rootdepth = 0.35}.
\item 1 to 5\% tree / shrub cover: \code{rootdepth = 0.50}.
\item 5 to 50\% tree / shrub cover: \code{rootdepth = 0.90}.
\item more than 50\% tree / shrub cover: \code{rootdepth = 1.10}.
}

The nomenclature of these objects is inspired by the sf package.
\itemize{
\item "sxc" stands for "spatial cross-section column", and like sf's "sfc"
objects, contains the cross section geometries.
\item "sx" stands for "spatial cross-section", and like sf's "sf" objects,
is a data frame where one of the columns is an "sxc" object, and
other columns are features of the cross section (roughness, d50, etc.)
}
}
\examples{
a <- xt_sxc(1:3)
xt_sx(1:3, swimability = c(4, 2, 1))
xt_sx(a, swimability = c(4, 2, 1))

}
